feat(core): refactor logger
